Restructurize the files
